updated example
